Moves localization code to its own separate project
Removed DotNetZip

+++ TombLauncher/Progress/CopyProgressInfo.cs
@@ -3,8 +3,8 @@
 public class CopyProgressInfo
 {
     public string CurrentFileName { get; set; }
-    public int TotalFiles { get; set; }
-    public int CurrentFile { get; set; }
+    public long TotalFiles { get; set; }
+    public long CurrentFile { get; set; }
     public double? Percentage => TotalFiles == 0 ? null : CurrentFile * 100 / TotalFiles;
     public string Message { get; set; }
 }
